+++ ChangeLog.md
@@ -2,6 +2,11 @@
 
 ## 0.2
 
+* Added `Reflex.Host.Basic.repeatUntilQuit_`, with the same behaviour
+  as 0.1's `Reflex.Host.Basic.repeatUntilQuit`.
+* `Reflex.Host.Basic.repeatUntilQuit` now returns an `Event` that
+  fires each time the action executes. If you don't need this,
+  consider `Reflex.Host.Basic.repeatUntilQuit_`.
 * `Reflex.Host.Basic.basicHostWithQuit`: Expect a guest that returns
   only `Event t ()`, as trying to return an actual result gives
   hard-to-diagnose type errors at the use site and most people